+++ pywikibot/families/vikidia_family.py
@@ -5,6 +5,7 @@
 # Distributed under the terms of the MIT license.
 #
 from pywikibot import family
+from pywikibot.tools import classproperty
 
 
 class Family(family.SubdomainFamily):
@@ -15,9 +16,30 @@ class Family(family.SubdomainFamily):
     domain = 'vikidia.org'
 
     codes = [
-        'ca', 'de', 'el', 'en', 'es', 'eu', 'fr', 'hy', 'it', 'oc', 'pt', 'ru',
-        'scn',
+        'ar', 'ca', 'de', 'el', 'en', 'es', 'eu', 'fr', 'hy', 'it', 'nl', 'oc',
+        'pt', 'ru', 'scn',
     ]
 
     # Sites we want to edit but not count as real languages
     test_codes = ['central', 'test']
+
+    @classproperty
+    def domains(cls):
+        """List of domains used by Vikidia family."""
+        return [
+            cls.domain,
+            'wikikids.nl'  # nl
+        ]
+
+    @classproperty
+    def langs(cls):
+        """Property listing family languages."""
+        cls.langs = super().langs
+        cls.langs['nl'] = cls.domains[1]
+        return cls.langs
+
+    def scriptpath(self, code):
+        """Return the script path for this family."""
+        if code == 'nl':
+            return ''
+        return super().scriptpath(code)