+++ tests/file_tests.py
@@ -13,9 +13,11 @@ from itertools import chain
 
 import pywikibot
 from pywikibot.exceptions import (
+    Error,
     NoPageError,
     NoWikibaseEntityError,
     PageRelatedError,
+    UserRightsError,
 )
 from pywikibot import pagegenerators
 from tests import join_images_path
@@ -351,7 +353,7 @@ class TestFilePageDataItem(TestCase):
         self.assertTrue(item.file is page)
         self.assertEqual('-1', item.id)
         item.get()
-        self.assertEqual('M14634781', item.getID())
+        self.assertEqual('M14634781', item.id)
         self.assertIsInstance(
             item.labels, pywikibot.page._collections.LanguageDict)
         self.assertIsInstance(
@@ -371,6 +373,15 @@ class TestFilePageDataItem(TestCase):
         del item._file
         self.assertEqual(page, item.file)
 
+    def test_data_item_not_file(self):
+        """Test data item with invalid pageid."""
+        item = pywikibot.MediaInfo(self.site, 'M1')  # Main Page
+        with self.assertRaises(Error):
+            item.file
+        with self.assertRaises(NoWikibaseEntityError):
+            item.get()
+        self.assertFalse(item.exists())
+
     def test_data_item_when_no_file_or_data_item(self):
         """Test data item associated to file that does not exist."""
         page = pywikibot.FilePage(self.site,
@@ -381,6 +392,8 @@ class TestFilePageDataItem(TestCase):
 
         with self.assertRaises(NoWikibaseEntityError):
             item.get()
+        with self.assertRaises(NoWikibaseEntityError):
+            item.labels
 
     def test_data_item_when_file_exist_but_without_item(self):
         """Test if data item is missing from file."""
@@ -393,25 +406,31 @@ class TestFilePageDataItem(TestCase):
         )
 
         # Seek to first page without mediainfo.
-        for page in gen:
-            if 'mediainfo' not in page.latest_revision.slots:
-                item = page.data_item()
-                self.assertIsInstance(item, pywikibot.MediaInfo)
-
-                # Get fails as there is no mediainfo.
-                with self.assertRaises(NoWikibaseEntityError):
-                    item.get()
-
-                # Create new empty mediainfo.
-                item.get_data_for_new_entity()
-                self.assertIsInstance(
-                    item.labels, pywikibot.page._collections.LanguageDict)
-                self.assertIsInstance(
-                    item.statements,
-                    pywikibot.page._collections.ClaimCollection)
-
-                # break the loop after checking first file
+        # Retry loop is for excepting incorrect files
+        for retry in range(5):
+            try:
+                for page in gen:
+                    slots = page.latest_revision.slots
+                    if 'mediainfo' not in slots:
+                        break
                 break
+            except ValueError:
+                pass
+
+        item = page.data_item()
+        self.assertIsInstance(item, pywikibot.MediaInfo)
+
+        # Get fails as there is no mediainfo.
+        with self.assertRaises(NoWikibaseEntityError):
+            item.get()
+
+        self.assertFalse(item.exists())
+        self.assertEqual(f'M{page.pageid}', item.id)
+        self.assertIsInstance(
+            item.labels, pywikibot.page._collections.LanguageDict)
+        self.assertIsInstance(
+            item.statements,
+            pywikibot.page._collections.ClaimCollection)
 
     def test_data_list_to_dict_workaround(self):
         """Test that T222159 workaround converts [] to {}."""
@@ -425,6 +444,112 @@ class TestFilePageDataItem(TestCase):
             item.labels, pywikibot.page._collections.LanguageDict)
         self.assertIsInstance(
             item.statements, pywikibot.page._collections.ClaimCollection)
+
+
+class TestMediaInfoReadonlyEditing(TestCase):
+
+    """Test writing structured data of FilePage."""
+
+    # Create valid API requests which will fail to missing permission.
+
+    # commons.wikimedia.beta.wmflabs.org
+    family = 'commons'
+    code = 'beta'
+
+    def test_edit_label(self):
+        """Test label editing without writing rights."""
+        page = pywikibot.FilePage(self.site, 'File:123_4.jpg')
+        item = page.data_item()
+        error_message = 'User "None" does not have required user right "edit"'
+
+        data = {}
+        lang = 'fi'
+        label = 'Edit label with editEntity and dict()'
+        data.update({'labels': {lang: label}})
+        with self.assertRaisesRegex(UserRightsError, error_message):
+            item.editEntity(data, summary=label)
+
+        data = item.get()
+        label = 'Edit label with editEntity and item.get()'
+        data.update({'labels': {lang: label}})
+        with self.assertRaisesRegex(UserRightsError, error_message):
+            item.editEntity(data, summary=label)
+
+        data = item.get(force=True)
+        label = 'Edit label with editEntity and item.get(force=True)'
+        data.update({'labels': {lang: label}})
+        with self.assertRaisesRegex(UserRightsError, error_message):
+            item.editEntity(data, summary=label)
+
+        lang = 'fi'
+        label = 'Edit label with editLabels and dict()'
+        data = {lang: label}
+        with self.assertRaisesRegex(UserRightsError, error_message):
+            item.editLabels(data, summary=label)
+
+        # Test label editing when file doesn't exists
+        page = pywikibot.FilePage(self.site, 'File:123_4_DOESNT_EXISTS.jpg')
+        item = page.data_item()
+
+        data = {}
+        lang = 'fi'
+        label = 'Edit label of missing file with editEntity and dict()'
+        data.update({'labels': {lang: label}})
+        with self.assertRaises(UserRightsError):
+            item.editEntity(data, summary=label)
+
+        lang = 'fi'
+        label = 'Edit label of missing file with editLabels and dict()'
+        data = {lang: label}
+        with self.assertRaisesRegex(UserRightsError, error_message):
+            item.editLabels(data, summary=label)
+
+    def test_edit_claims(self):
+        """Test claim editing without writing rights."""
+        error_message = 'User "None" does not have required user right "edit"'
+        wikidata_site = pywikibot.Site('wikidata', 'wikidata')
+
+        # Test adding claim existing file
+        page = pywikibot.FilePage(self.site, 'File:123_4.jpg')
+        item = page.data_item()
+
+        # Create claim
+        finna_id = 'test123'
+        new_claim = pywikibot.Claim(wikidata_site, 'P9478')
+        new_claim.setTarget(finna_id)
+
+        # Do actual edit
+        with self.assertRaisesRegex(UserRightsError, error_message):
+            item.addClaim(new_claim)
+
+        # Test removing first claim using item.statements
+        for property_id in item.statements:
+            for statement in item.statements[property_id]:
+                with self.assertRaisesRegex(UserRightsError, error_message):
+                    summary = f'Removing {property_id}'
+                    item.removeClaims(statement, summary=summary)
+                break
+
+        # Test removing first claim using item.claims
+        for property_id in item.claims:
+            for claim in item.claims[property_id]:
+                with self.assertRaisesRegex(UserRightsError, error_message):
+                    summary = f'Removing {property_id}'
+                    item.removeClaims(claim, summary=summary)
+                break
+
+        # Test adding claim to non-existing file
+        page = pywikibot.FilePage(self.site, 'File:123_4_DOESNT_EXISTS.jpg')
+        item = page.data_item()
+
+        # Create claim
+        finna_id = 'test123'
+        new_claim = pywikibot.Claim(wikidata_site, 'P9478')
+        new_claim.setTarget(finna_id)
+
+        # Do actual edit
+        with self.assertRaises(NoWikibaseEntityError):
+            item.addClaim(new_claim)
 
 
 if __name__ == '__main__':  # pragma: no cover